pyproject.toml
[project]
name = "okay-color-names"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = [
    "matplotlib>=3.9.2",
    "networkx>=3.3",
    "tqdm>=4.66.5",
    "scikit-learn>=1.5.2",
    "Pillow>=10.4.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "GPL-3.0"}


[tool.pdm]
distribution = true

[tool.pdm.scripts]
this = "python -m okay_color_names"
